Set up A-MoreOOP
